+++ apps/crates.io-parser/src/config/logger.ts
@@ -1,5 +1,0 @@
-import pino from "pino";
-
-const _logger = pino();
-
-export const logger = _logger;