+++ apps/crates.io-parser/src/types.ts
@@ -1,0 +1,6 @@
+import type { S3Client } from "@aws-sdk/client-s3";
+
+export type AppState = {
+  minioClient: S3Client;
+  bucket: string;
+};