+++ apps/crates.io/src/main.ts
@@ -1,21 +1,13 @@
-import { HttpClient } from "@scrappers/http-client";
-import amqplib from "amqplib";
-import { Telemetry } from "@scrappers/telemetry";
-import { StorageMinio } from "@scrappers/storage-minio";
-import type { Storage } from "@scrappers/storage";
-import { logger } from "./config/logger";
-import { trace } from "@opentelemetry/api";
+import { Telemetry } from "@packages/telemetry";
 import dotenv from "dotenv";
-import { Config } from "./config/config";
-import { env } from "./types/env";
-
-const QUEUE = "integration.crates.io";
-const tracer = trace.getTracer("crates.io");
+import { Config } from "@packages/config";
+import { CONFIG_DIR, CONFIG_SCHEMA } from "./config";
+import { App } from "./app";
 
 const main = async (): Promise<void> => {
   dotenv.config();
 
-  const config = await Config(env());
+  const config = await Config.build(CONFIG_DIR, CONFIG_SCHEMA);
 
   const telemetry = Telemetry("crates.io");
 
@@ -24,92 +16,12 @@ const main = async (): Promise<void> => {
     serviceName: config.name,
   });
 
-  const storage = StorageMinio({
-    bucket: "integrations",
-    endpoint: config.minio.url,
-    credentials: {
-      accessKeyId: config.minio.username,
-      secretAccessKey: config.minio.password,
-    },
-    region: config.minio.region,
-    forcePathStyle: config.minio.force_path_style,
-  });
+  const app = await App.build(config);
 
-  const connection = await amqplib.connect(config.rabbitmq.url);
-  const channel = await connection.createChannel();
-  await channel.assertQueue(QUEUE, {
-    durable: true,
-  });
-
-  logger.info("Connected to RabbitMQ");
-
-  const client = new HttpClient("https://crates.io");
-
-  await channel.consume(QUEUE, async (message) => {
-    if (!message) return;
-
-    const headers = message.properties.headers || {};
-
-    const context = telemetry.context({
-      traceparent: headers.traceparent,
-      tracestate: headers.tracestate,
-    });
-
-    await telemetry.propagate(context, async () => {
-      await telemetry.span("start-scraping", async () => {
-        await consume(message, channel, client, storage);
-      });
-    });
-
-    const forwardCarrier = telemetry.carrier(context);
-
-    channel.publish(
-      "default_exchange",
-      "integration.crates.io.parser",
-      message.content,
-      {
-        headers: {
-          traceparent: forwardCarrier.traceparent,
-          tracestate: forwardCarrier.tracestate,
-        },
-        contentType: "application/json",
-      },
-    );
-  });
+  await app.run();
 };
 
 main().catch((error) => {
   console.error("Fatal error:", error);
   process.exit(1);
 });
-
-const consume = async (
-  message: amqplib.Message,
-  channel: amqplib.Channel,
-  client: HttpClient,
-  storage: Storage,
-): Promise<void> => {
-  const payload = JSON.parse(message.content.toString());
-
-  try {
-    const response = await client.request({
-      method: "GET",
-      path: `/api/v1/crates/${payload.package_name}`,
-      headers: {
-        "User-Agent":
-          "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/123.0.0.0 Safari/537.36",
-      },
-    });
-
-    const content = await response.body.json();
-
-    const span = tracer.startSpan("write-to-storage");
-    await storage.put(`crates/${payload.package_name}.json`, content);
-    span.end();
-
-    channel.ack(message);
-  } catch (error) {
-    logger.info("Error consuming message", { error });
-    channel.nack(message, false, false);
-  }
-};