+++ packages/core/src/index.ts
@@ -1,0 +1,3 @@
+export * from "@packages/config";
+export * from "@packages/env";
+export * from "@packages/telemetry";