+++ packages/telemetry/src/functions/context.ts
@@ -1,5 +1,6 @@
+import type { Context } from "@opentelemetry/api";
 import { context as otelContext, propagation } from "@opentelemetry/api";
 
-export const context = (carrier: Record<string, string> = {}) => {
+export const context = (carrier: Record<string, string> = {}): Context => {
   return propagation.extract(otelContext.active(), carrier);
 };