+++ packages/telemetry/src/index.ts
@@ -4,7 +4,15 @@ import { init } from "./functions/init";
 import { propagate } from "./functions/propagate";
 import { span } from "./functions/span";
 
-export const Telemetry = (serviceName: string) => ({
+export type Telemetry = {
+  init: typeof init;
+  context: typeof context;
+  propagate: typeof propagate;
+  span: ReturnType<typeof span>;
+  carrier: typeof carrier;
+};
+
+export const Telemetry = (serviceName: string): Telemetry => ({
   init,
   context,
   propagate,